3rd day authentication and many libraries

+++ server/src/routes/memories.ts
@@ -8,9 +8,18 @@ import { z } from 'zod'
 
 // Aqui vai todas as rotas voltadas a parte de gerenciamento de memórias
 export async function memoriesRoutes(app: FastifyInstance) {
+  // Antes de executar o Handler (as funções abaixo) de cada uma das rotas, verifique se o usuário está autenticado
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify() // Garante que o Token está vindo na requisição
+  })
+
   // READ
-  app.get('/memories', async () => {
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      // Retorna somente as memórias do usar logado
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc', // Ordenar por datas da mais antiga para mais nova
       },
@@ -25,7 +34,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
       }
     })
   })
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -38,6 +47,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id, // id: id,
       },
     })
+
+    // Se a memória não for pública e o ID dessa memória for diferente do usuário logado, retornar erro 401
+    if (!memory.isPublic && memory.userId !== request.user.sub)
+      return reply.status(401).send()
 
     return memory // TODO: Tratar o erro 500 caso não seja encontrado o Id
   })
@@ -56,7 +69,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '493ac551-c7ab-4940-b5d4-bb03464a27e3', // Estático por hora
+        userId: request.user.sub,
       },
     })
 
@@ -64,7 +77,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
   })
 
   // UPDATE
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -79,7 +92,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) return reply.status(401).send()
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -94,12 +115,20 @@ export async function memoriesRoutes(app: FastifyInstance) {
   })
 
   // DELETE
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) return reply.status(401).send()
 
     await prisma.memory.delete({
       where: {
